Feat: 유저 유효성 검사

See merge request s11-bigdata-dist-sub1/S11P21B206!77

+++ backend/user-service/src/main/java/com/helloword/userservice/domain/user/service/UserServiceImpl.java
@@ -12,6 +12,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.regex.Pattern;
+
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
@@ -19,9 +21,21 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder customPasswordEncoder;
 
+    private static final String EMAIL_PATTERN = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$";
+    private static final String PHONE_PATTERN = "^010-\\d{4}-\\d{4}$";
+
     @Override
     public void registerUser(SignupRequest signupRequest) {
         signupRequest.setPassword(customPasswordEncoder.encode(signupRequest.getPassword()));
+
+        if (!Pattern.matches(EMAIL_PATTERN, signupRequest.getEmail())) {
+            throw new ExceptionResponse(CustomException.INVALID_EMAIL_FORMAT);
+        }
+
+        if (!Pattern.matches(PHONE_PATTERN, signupRequest.getPhone())) {
+            throw new ExceptionResponse(CustomException.INVALID_PHONE_FORMAT);
+        }
+
         userRepository.save(User.createUser(signupRequest));
     }
 
@@ -36,15 +50,15 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Long userId) {
+    public void deleteUser(Long xUserId, Long userId) {
+        checkOwnership(xUserId, userId);
         userRepository.deleteById(userId);
     }
 
     @Override
-    public void updatePassword(Long userId, NewPasswordRequest newPasswordRequest) {
-        User user = userRepository.findById(userId).orElseThrow(
-                () -> new ExceptionResponse(CustomException.NOT_FOUND)
-        );
+    public void updatePassword(Long xUserId, Long userId, NewPasswordRequest newPasswordRequest) {
+        User user = findUserById(userId);
+        checkOwnership(xUserId, userId);
 
         user.changePassword(customPasswordEncoder.encode(newPasswordRequest.getPassword()));
 
@@ -62,5 +76,16 @@ public class UserServiceImpl implements UserService {
         }
 
         return AuthenticateUserResponse.toDto(user);
+    }
+
+    private User findUserById(Long userId) {
+        return userRepository.findById(userId).orElseThrow(
+                () -> new ExceptionResponse(CustomException.NOT_FOUND));
+    }
+
+    private void checkOwnership(Long ownerId, Long requestUserId) {
+        if (!ownerId.equals(requestUserId)) {
+            throw new ExceptionResponse(CustomException.FORBIDDEN);
+        }
     }
 }
